Combine styles
